Always preserve prov's config.yaml (#75605)
* Always preserve prov's config.yaml

Also, if for whatever reason it's not there, make sure to not overwrite
the one from write-config step.

A grace period needs to be explicitly defined for this to work.

* Add some context on these files

+++ ci-operator/step-registry/aro-hcp/provision/environment/aro-hcp-provision-environment-commands.sh
@@ -62,15 +62,29 @@ yq eval -n "
 echo "Created override config at: ${OVERRIDE_CONFIG_FILE}"
 cat ${OVERRIDE_CONFIG_FILE}
 
+CONFIG_PROV="${SHARED_DIR}/config-prov.yaml"
+
+# There's a $SHARED_DIR/config.yaml already from the write-config step
+# but it is of limited accuracy. It's fine for int/stg/prod, but this prov
+# step will generate temporary names for a bunch of things, so if we want
+# following steps to know what those are, we need to override the older
+# less accurate config.yaml.
+# And let's do it in a way that works even if provisioning ends up failing.
+finalize() {
+    if [[ -s "${CONFIG_PROV}" ]]; then
+        mv "${CONFIG_PROV}" "${SHARED_DIR}/config.yaml"
+        cp "${SHARED_DIR}/config.yaml" "${ARTIFACT_DIR}/config.yaml"
+    fi
+}
+trap finalize EXIT
+
 unset GOFLAGS
 make -o tooling/templatize/templatize entrypoint/Region \
   DEPLOY_ENV=prow \
   EXTRA_ARGS="--region ${LOCATION} --abort-if-regional-exist" \
   TIMING_OUTPUT=${SHARED_DIR}/steps.yaml.gz \
   ENTRYPOINT_JUNIT_OUTPUT=${ARTIFACT_DIR}/junit_entrypoint.xml \
-  CONFIG_OUTPUT=${SHARED_DIR}/config.yaml
-
-cp "${SHARED_DIR}/config.yaml" "${ARTIFACT_DIR}/config.yaml"
+  CONFIG_OUTPUT=${CONFIG_PROV}
 
 # Mark successful completion
 touch "${SHARED_DIR}/provision-complete"
